fix/servico_delete

+++ src/Http/Controllers/Servico/ServicoController.php
@@ -199,17 +199,27 @@ class ServicoController extends Controller
     public function deleteServico(Request $request, $servicoId)
     {
         $userId = Auth::id($request);
+        $servicoUser = $this->servico->getServicoUser($servicoId);
         
-        try
+        if($servicoUser != $userId)
         {
-            $this->servico->deleteServico($servicoId);
-            return Response::json(['message' => 'Serviço deletado com sucesso']);
+            return Response::json(['message' => 'Você não pode deletar o serviço de outro usuário']);
         }
-        catch(DatabaseInsertException $exception)
+        else
         {
-            DB::rollback();
-            return Response::json(['error' => ['message' => $exception->getMessage()]], $exception->getCode());
+            try
+            {
+                $this->servico->deleteServico($servicoId);
+                return Response::json(['message' => 'Serviço deletado com sucesso']);
+            }
+            catch(DatabaseInsertException $exception)
+            {
+                DB::rollback();
+                return Response::json(['error' => ['message' => $exception->getMessage()]], $exception->getCode());
+            }
         }
+
+        
     }
     
     public function uploadImage(Request $request, $servicoId)
